Update jackpot.c to V0.3
added an check if the wallet is still charged so the player could not bet more money then he has and go into negative value. 
game telle the player he lost all the money and he lost the game or if he still has some money but tried to bet too much.